Thymus extra functions and analysis
Functions and procedures to the calculation of size musculus ensembl_gene_id
Analysis of methylation data from thymus DNMT1 KO.
Modification of hierarchies and parameters for the deconvolution of thymus data. Addition of partial reference por thymus.
Complete notebook using Decoflex to deconvolute thymus data in the DNMT1 KO with a different single-cell data. Also a partial or semi-reference model is created. Finally a similar approach was implemente with Hdac1 KO noting that the WT dataset was not compatible with Elisa data's paper.